requirements.txt.txt
torch==2.0.1
torchvision==0.15.2
tensorflow==2.12.0
streamlit==1.25.0
Pillow==10.0.0
numpy==1.25.2
matplotlib==3.8.2
scikit-learn==1.3.1
kagglehub==1.3.0